ImageComparator/requirements.txt
opencv-python==4.8.1.78
dhash==1.4
Pillow==10.1.0